some fixes

+++ handler/quit.go
@@ -3,10 +3,8 @@ package handler
 import "github.com/df-mc/dragonfly/server/entity/damage"
 
 func (h PlayerHandler) HandleQuit() {
-	if !h.p.CombatCD().Expired() {
-		if src, ok := h.p.LastHurt(); ok {
-			h.p.Kill(damage.SourceEntityAttack{Attacker: src})
-		}
+	if src, ok := h.p.LastHurt(); ok && !h.p.CombatCD().Expired() {
+		h.p.Kill(damage.SourceEntityAttack{Attacker: src})
 	}
 	h.p.Server().RemovePlayer(h.p)
 }
